refactor: change the hacky redirection to callback
On the creation of a new todo, the item has to be added to store. Then the effect has to wait till the length of todo list changes. Then redirect to the last one in the list. This is hacky.

This change will wrap the redirection logic inside a callback. Instead of hacky sideeffect.